Added Projects (Read Description)
Added the following projects:
*CUDA Vector Interface
*Gap Learning Timer
*Taxbot-9000
*Text-to-GPT
*Valorant Stat Collection and Format
*Voice to Voice Translator

If you are going to check any of these projects out, the ones I put the most effort into are the Valorant, Translator, and CUDA Interface projects. I recommend taking a look at the CUDA project as it is the most practical, and is the beginning of a custom neural network I am working on.